# dbconf/swagger3.yaml
swagger: "2.0"
info:
  title: Ecomerse Api
  description: API.
  version: 1.0.0
host: localhost:7000
basePath: /api
schemes:
  - http
paths:
...
  /user/single:
    get:
      tags:
        - "user"
      summary: "Find single user"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findAdminByData"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "query"
          description: "5dd66a92f3a23a1b4065fa2a"
          required: true
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#"
        400:
          description: "Invalid status value"
    post:
      tags:
        - "user"
      summary: "Add User in the store with form data"
      description: ""
      operationId: "updatePetWithForm"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "firstName"
          in: "formData"
          description: "add name of the User"
          required: false
          type: "string"
        - name: "lastName"
          in: "formData"
          description: "add name of the User"
          required: false
          type: "string"
        - name: "email"
          in: "formData"
          description: "add email of the User"
          required: false
          type: "string"
        - name: "phone"
          in: formData
          description: "add email of the User"
          required: false
          type: "string"
        - name: "password"
          in: "formData"
          description: "add password of the User"
          required: false
          type: "string"
        - name: "role"
          in: "formData"
          description: "role"
          required: false
          type: "string"
      responses:
        405:
          description: "Invalid input"
    put:
      tags:
        - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
        - "application/json"
      consumes:
        - "application/x-www-form-urlencoded"
      parameters:
        - name: "id"
          in: "query"
          description: "5dde41fd7ac7e72f14479b69"
          required: true
          type: "string"
        - name: "firstName"
          in: "formData"
          description: "add name of the User"
          required: false
          type: "string"
        - name: "lastName"
          in: "formData"
          description: "add name of the User"
          required: false
          type: "string"
        - name: "email"
          in: "formData"
          description: "add email of the User"
          required: false
          type: "string"
        - name: "phone"
          in: formData
          description: "add email of the User"
          required: false
          type: "string"
        - name: "password"
          in: "formData"
          description: "add password of the User"
          required: false
          type: "string"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
    delete:
      tags:
        - "user"
      summary: "Deletes a User"
      description: ""
      operationId: "deleteUser"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "query"
          description: "delete - user - id"
          required: true
          type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
...